@@ -11,11 +11,11 @@
     "@testing-library/user-event": "^14.6.1",
     "@unocss/reset": "^0.53.6",
     "axios": "^1.8.3",
-    "react": "^18.3.1",
+    "react": "^19.0.0",
     "react-cookie": "^8.0.1",
-    "react-dom": "^18.3.1",
+    "react-dom": "^19.0.0",
     "react-redux": "^9.2.0",
-    "react-router-dom": "^5.3.0"
+    "react-router-dom": "^7.3.0"
   },
   "devDependencies": {
     "@vitejs/plugin-react-swc": "^3.3.2",
