@@ -1,15 +1,15 @@
-import { useDispatch } from 'react-redux'
 import { useCallback } from 'react'
-import { useHistory } from 'react-router-dom'
+import { useDispatch } from 'react-redux'
+import { useNavigate } from 'react-router-dom'
 import { logout } from '~/store/auth'
 
 export const useLogout = () => {
   const dispatch = useDispatch()
-  const history = useHistory()
+  const navigate = useNavigate()
 
   const handleLogout = useCallback(async () => {
     await dispatch(logout()).unwrap()
-    history.push('/signin')
+    navigate('/signin')
   }, [useDispatch])
 
   return {
