@@ -1,25 +1,16 @@
-import { useDispatch } from 'react-redux'
 import { useCallback } from 'react'
-import { useHistory } from 'react-router-dom'
+import { useDispatch } from 'react-redux'
+import { useNavigate } from 'react-router-dom'
 import { signup } from '~/store/auth'
 
 export const useSignup = () => {
   const dispatch = useDispatch()
-  const history = useHistory()
+  const navigate = useNavigate()
 
-  const handleSignup = useCallback(
-    async ({ email, password, name }) => {
-      await dispatch(
-        signup({
-          email,
-          password,
-          name,
-        }),
-      ).unwrap()
-      history.push('/')
-    },
-    [useDispatch],
-  )
+  const handleSignup = useCallback(async (email, password) => {
+    await dispatch(signup({ email, password })).unwrap()
+    navigate('/')
+  }, [useDispatch])
 
   return {
     signup: handleSignup,
