@@ -1,10 +1,10 @@
-import React, { StrictMode } from 'react'
-import ReactDOM from 'react-dom'
-import './index.css'
-import App from './App'
+import { StrictMode } from 'react'
+import { createRoot } from 'react-dom/client'
 import { Provider } from 'react-redux'
-import { store } from './store/index'
 import axios from '~/vendor/axios'
+import App from './App'
+import './index.css'
+import { store } from './store/index'
 
 axios.interceptors.request.use(config => {
   const token = store.getState().auth.token
@@ -16,11 +16,11 @@ axios.interceptors.request.use(config => {
 })
 
 const root = document.getElementById('root')
-ReactDOM.render(
+const rootElement = createRoot(root)
+rootElement.render(
   <StrictMode>
     <Provider store={store}>
       <App />
     </Provider>
-  </StrictMode>,
-  root,
+  </StrictMode>
 )
