@@ -1,16 +1,16 @@
-import { useCallback, useState, useEffect } from 'react'
-import { Link, useHistory, useParams } from 'react-router-dom'
-import { useSelector, useDispatch } from 'react-redux'
+import { useCallback, useEffect, useState } from 'react'
+import { useDispatch, useSelector } from 'react-redux'
+import { Link, useNavigate, useParams } from 'react-router-dom'
 import { BackButton } from '~/components/BackButton'
-import './index.css'
-import { fetchLists, updateList, deleteList } from '~/store/list'
 import { useId } from '~/hooks/useId'
+import { deleteList, fetchLists, updateList } from '~/store/list'
+import './index.css'
 
 const EditList = () => {
   const id = useId()
 
   const { listId } = useParams()
-  const history = useHistory()
+  const navigate = useNavigate()
   const dispatch = useDispatch()
 
   const [title, setTitle] = useState('')
@@ -41,7 +41,7 @@ const EditList = () => {
       void dispatch(updateList({ id: listId, title }))
         .unwrap()
         .then(() => {
-          history.push(`/lists/${listId}`)
+          navigate(`/lists/${listId}`)
         })
         .catch(err => {
           setErrorMessage(err.message)
@@ -63,7 +63,7 @@ const EditList = () => {
     void dispatch(deleteList({ id: listId }))
       .unwrap()
       .then(() => {
-        history.push(`/`)
+        navigate(`/`)
       })
       .catch(err => {
         setErrorMessage(err.message)
