@@ -1,8 +1,8 @@
-import React, { useCallback, useState } from 'react'
-import { Redirect, Link } from 'react-router-dom'
-import { useSelector } from 'react-redux'
-import { useLogin } from '~/hooks/useLogin'
-import { useId } from '~/hooks/useId'
+import {useCallback, useState} from 'react'
+import {useSelector} from 'react-redux'
+import {Link, Navigate} from 'react-router-dom'
+import {useId} from '~/hooks/useId'
+import {useLogin} from '~/hooks/useLogin'
 import './index.css'
 
 const SignIn = () => {
@@ -34,7 +34,7 @@ const SignIn = () => {
   )
 
   if (auth) {
-    return <Redirect to="/" />
+    return <Navigate to="/" />
   }
 
   return (
